Rename prod to publish tsconfig json file
Re-organize exports for server/indexer exports
