+++ src/typescript/sdk/src/client/emojicoin-client.ts
@@ -21,7 +21,7 @@ import {
 } from "../emojicoin_dot_fun/emojicoin-dot-fun";
 import { type Events } from "../emojicoin_dot_fun/events";
 import { getEmojicoinMarketAddressAndTypeTags } from "../markets";
-import { type EventsModels, getEventsAsProcessorModelsFromResponse } from "../mini-processor";
+import { type EventsModels, getEventsAsProcessorModelsFromResponse } from "../indexer-v2";
 import { APTOS_CONFIG, getAptosClient } from "../utils/aptos-client";
 import { toChatMessageEntryFunctionArgs } from "../emoji_data";
 import customExpect from "./expect";
