+++ src/typescript/sdk/src/index.ts
@@ -1,10 +1,6 @@
 export * from "./const";
-export * from "./client";
 export * from "./emoji_data";
 export * from "./emojicoin_dot_fun";
 export * from "./markets";
-export * from "./mini-processor";
-export * from "./queries";
-export * from "./server";
 export * from "./types";
 export * from "./utils";
