+++ src/typescript/sdk/src/indexer-v2/index.ts
@@ -1,3 +1,4 @@
 export * from "./json-bigint";
+export * from "./mini-processor";
 export * from "./queries";
 export * from "./types";
