+++ src/typescript/sdk/src/indexer-v2/mini-processor/event-groups/utils.ts
